refactor: consolidate SettingsData and migrate MCP to TypeScript
- Replace 6 duplicate SettingsData interfaces with shared ServerSettings from types/config.ts
- Migrate MCP module (493 lines) from JavaScript to TypeScript
- Add proper typing for MCP tool responses, artwork results, and API helpers
- All 220 tests passing
